move util from model to cfn_utils
